bot/requirements.txt
aiogram
requests
python-dotenv

